Rename getConfig to getDataConfig
