+++ packages/core/src/js/utilities/getDataConfig.js
@@ -1,4 +1,4 @@
-export function getConfig(el, dataConfig = "config") {
+export function getDataConfig(el, dataConfig = "config") {
   const string = el.getAttribute(`data-${dataConfig}`) || "";
   const json = string.replace(/'/g, "\"");
   return (json) ? JSON.parse(json) : {};
